feat(ui): add horizontal module subnav below dashboard header
each module area now shows a sticky horizontal tab bar with only its
own sub-routes. the active tab is highlighted with the module accent
color and an underline indicator.

- extract navigation config to src/constants/navigation.ts
- new ModuleSubnav component reads pathname and renders relevant tabs
- sidebar imports from constants instead of duplicating data
- fix user.service createUser to pass companyId for prisma types
